trunk/nitrate/requirements/devel.txt
-r base.txt
django-debug-toolbar==0.11.0
Sphinx==1.2b3
